Reorganize metal to support 2WB in Glimmer
This commit restructures some of the more arcane parts of ember-metal's
observer system.

It eliminates some special casing around `length`, which isn't necessary
anymore but complicates the overall abstraction of observable
properties.

Relatedly, it also finishes implementing PROPERTY_DID_CHANGE, which now
executes reliably any time an observer would have fired. You can think
of it as an always-on observer if it's present on an object. Glimmer 2
curly components use this hook to propagate `set`s on components back to
their original bindings (if the bindings are present).

This commit also implements two way bindings in Glimmer 2, by making
PropertyReferences implement an `UPDATE` method. If that method is
present, it is used to propagate changes upwards (otherwise, an
exception is thrown).